backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.13.1
pydantic==2.5.0
python-multipart==0.0.6
PyMuPDF==1.23.8
python-dateutil==2.8.2
httpx==0.25.2